// src/app/demos/directory/components/types.ts
export type StatType = {
  image: string
  name: string
  count: number
}

export type DirectoryType = {
  title: string
  description: string
  price: number
  avatar: string
  name: string
  image: string
  rating: number
}

export type stepType = {
  icon: string
  title: string
  description: string
}

export type testimonialType = {
  avatar: string
  name: string
  position: string
  description: string
}

export type BlogType = {
  image: string
  date: string
  title: string
  description: string
}